security(deps): clear all RustSec advisories and gate on cargo-audit
Closes #584.

`cargo audit` went from 11 vulnerabilities and 7 warnings across 587 crates
to zero of either across 563. Most of it was `cargo update` — the advisories
were overwhelmingly transitive: crossbeam-epoch, postgres-protocol,
tokio-postgres, rustls-webpki and quinn-proto all had patch releases waiting.

Two needed a manifest change. `prometheus` gets `default-features = false`,
dropping its `protobuf` feature and with it protobuf 2.28.0
and RUSTSEC-2024-0437 — admin.rs only ever uses `TextEncoder`. And
`object_store` goes 0.11.2 to 0.14.1, the first release built on
quick-xml >= 0.41; it was the last thing pulling the version affected by
RUSTSEC-2026-0194/0195 after the three direct dependants moved to 0.41. That
also collapses two parallel object_store trees into one, since `icechunk`
already linked 0.14.

The quick-xml major is behavioural, not cosmetic. Since 0.38 an entity
reference terminates the surrounding text event and arrives as its own
`Event::GeneralRef`. Both CAP readers matched only `Event::Text`, so they
would have silently dropped every `&`, `<` and `>` from alert headlines,
descriptions and feed URLs — silently, because the document stays
well-formed and nothing errors. `parser::push_general_ref` handles the new
event, and the reader-side `trim_text(true)` had to go with it: per-event
trimming eats the spaces around an entity, turning `Wind &amp; Flood` into
`Wind&Flood`. Every leaf was already trimmed whole at its closing tag.

object_store 0.14 also moved `get`/`head`/`get_range` into an
`ObjectStoreExt` extension trait, widened `ObjectMeta::size` to `u64`, and
takes `Range<u64>` for `get_range`. `DataStore` keeps its `Range<usize>`
signature so no caller changes.

The gate: a new `audit` job runs on PRs, on main, and on the weekly
schedule — the schedule being the point, since an advisory published against
a crate we already ship turns CI red with nobody touching the code. On main
it also gates the Docker publish, so the weekly base-image refresh cannot
republish `:main` on a vulnerable tree. Policy lives in `.cargo/audit.toml`
rather than workflow flags so a bare `cargo audit` locally gives the
identical verdict, and it denies warnings — unmaintained, unsound and yanked
crates fail too, since letting those slide is how #584 accumulated in the
first place. `cargo metadata --locked` runs first because a lockfile that
disagrees with the manifests would make the audit meaningless.

`paste` and `lru` stay ignored with reachability analysis in that file;
neither is fixable from here. Tracked in #594.

Dependabot covers direct crates and GitHub Actions weekly, grouped and
capped at 3 open PRs.

+++ crates/engine-geotiff/src/catalog.rs
@@ -528,7 +528,9 @@ fn is_excluded(filename: &str, patterns: &[String]) -> bool {
 }
 
 /// Maximum file size for remote downloads (50 MB).
-pub(crate) const MAX_REMOTE_FILE_SIZE: usize = 50 * 1024 * 1024;
+/// `u64` to match `ObjectMeta::size`, which object_store widened from `usize`
+/// in 0.14 so 32-bit targets can still describe large objects.
+pub(crate) const MAX_REMOTE_FILE_SIZE: u64 = 50 * 1024 * 1024;
 
 /// Scan a remote object store for GeoTIFF files matching a filename pattern.
 ///
@@ -563,7 +565,7 @@ pub fn scan_remote_with_limit(
                 }
                 let key = obj.location.to_string();
                 let filename = key.rsplit('/').next().unwrap_or(&key).to_string();
-                Some((key, filename, obj.size as u64, obj.location.clone()))
+                Some((key, filename, obj.size, obj.location.clone()))
             })
             .collect();
